Implement authentication flow with token handling and UI updates
